refactor(desktop): split developer log export helpers
